+++ Plugins/Experimental/Experimentals/UnhardcodeRangerDualWield.hpp
@@ -1,0 +1,16 @@
+#pragma once
+
+#include "Common.hpp"
+
+namespace Experimental {
+
+class UnhardcodeRangerDualWield
+{
+public:
+    UnhardcodeRangerDualWield(NWNXLib::Services::HooksProxy* hooker);
+
+private:
+    static int32_t HasFeatHook(CNWSCreatureStats*, uint16_t);
+};
+
+}