+++ Plugins/Tweaks/Tweaks.cpp
@@ -22,6 +22,7 @@
 #include "Tweaks/FixArmorDexBonusUnderOne.hpp"
 #include "Tweaks/FixItemNullptrInCItemRepository.hpp"
 #include "Tweaks/ClearSpellEffectsOnTURDs.hpp"
+#include "Tweaks/AlwaysReturnFullDEXStat.hpp"
 
 #include "Services/Config/Config.hpp"
 
@@ -189,6 +190,12 @@ Tweaks::Tweaks(Services::ProxyServiceList* services)
     {
         LOG_INFO("Effects on logged out users will be removed when a caster rests.");
         m_ClearSpellEffectsOnTURDs = std::make_unique<ClearSpellEffectsOnTURDs>(GetServices()->m_hooks.get());
+    }
+
+    if (GetServices()->m_config->Get<bool>("ALWAYS_RETURN_FULL_DEX_STAT", false))
+    {
+        LOG_INFO("GetDEXStat() is always returning a creature's full Dexterity Stat.");
+        m_AlwaysReturnFullDEXStat = std::make_unique<AlwaysReturnFullDEXStat>(GetServices()->m_hooks.get());
     }
 }